src/main/java/com/devops/service/UserService.java
package com.devops.service;




import org.springframework.stereotype.Service;

import com.devops.dto.UserDto;
import com.devops.entity.User;
import com.devops.repository.UserRepository;

@Service
public class UserService {

    private final UserRepository userRepository;

    public UserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public User saveUser(UserDto userDto) {
        User user = new User();
        user.setName(userDto.getName());
        user.setEmail(userDto.getEmail());
        return userRepository.save(user);
    }
}
